// src/app/features/animal-health/first-aid/models/saveFirstAidDetails.model.ts
export interface SaveFirstAid {
    animalId:               number;
    ownerId:                number;
    tagId:                  number;
    treatmentType:          string;
    caseStatus:             string;
    ailmentCd:              number;
    firstAidTreatmentGiven: string;
    treatmentRemarks:       string;
    paymentAmount:          number;
    receiptNo:              string;
    treatmentRecordDate:    string;
    treatmentDate:          string;
    modifiedDate:           string;
    modifiedBy:             string;
    createdBy:              string;
    creationDate:           string;
}